Add a test exposing the issue
Only apply capped rollback items if the ns is still open. We know it
can only be closed/reopend if the same txn did fileops, so it is safe
to assume that if the ns is open, this txn did not do fileops and so
the in-memory stats must be fixed. Otherwise, this txn did do fileops
and thus the ns is closed because of the nsindex rollback items applied
beforehand.

Reuse less code for the sake of clarity. We should not have to call
init() and check allocated() again because we would have already
returned if the nsdb wasn't loaded.