added reporting of progress with respect to source #333
Added a new option --reportingPeriod which specifies how frequently to
report the current syncing progress, in seconds (10 is default).  Also
reports periodically when idle (whenever the tailing query times out).

This reports the current OpTime we're synced to, and does a query on the
source oplog to find out how far it has left to go, and reports how many
seconds behind the TokuMX cluster is.  This gives a good sense for
whether it's catching up over time.