fixed authorization for txn and loader commands #698
Previously, txn commands required authorization to be done on SERVER_RESOURCE_NAME, which means you
must be readAnyDatabase/readWriteAnyDatabase in order to do it.  This is unintuitive, we should
require very little permissions for transactions and let the constituent operations' authorization
checks do what they need to.  We don't want unauthorized users being able to run these commands
though, because that may allow for DoS, so we require at least the read role on whatever database
the command is run against (even though it doesn't matter which database you choose).

Similarly, the loader commands should require the readWrite role, not the readWriteAnyDatabase role.

fixes #698