hopefully fixed loader_begin.js #920
The theory is that the count is running while the parallel shell is still
running.  Normally the parallel shell should not run long enough for the
count to time out (4s) but it seems that on a heavily loaded system, we
may get scheduled poorly.  If this theory is true, then joining the
parallel shell before running the count should be sufficient to stop it
from failing intermittently.